change type of id to Long

+++ src/main/java/ua/shpp/dto/BranchRequestDTO.java
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.Size;
 @Schema(description = "Request to branch")
 public record BranchRequestDTO(
         @Schema(description = "Id of organization", example = "1")
-        Integer organizationId,
+        Long organizationId,
         @Schema(description = "Name of the branch", example = "Central branch")
         @Size(min = 6, max = 30, message = "The length of the branch should be between 6-30 characters")
         @NotBlank
